Se anadio el AsNoTracking en metodo de UpdatePartial para poder guardar los cambios

+++ MagicVilla_API/Models/DTO/VillaDto.cs
@@ -22,7 +22,7 @@ namespace MagicVilla_API.Models.Dto
 
         public int Ocupantes { get; set; }
 
-        public int MetrosCuadrados { get; set; }
+        public Double MetrosCuadrados { get; set; }
 
         public string ImagenUrl { get; set; }
         public string Amenidad { get; set; }
